fix: Profile page works after reloading
