+++ src/views/MyProfileView.vue
@@ -1,18 +1,15 @@
 <template>
     <MyProfile :user="loggedUser" />
     <CreateFlit />
-    <FlitFeed :posts="posts" 
-    @previousPage="previousPage()"
-    @nextPage="nextPage()"/>
+    <UserFlitFeed :user="loggedUser" />
 </template>
 
 <script lang="ts">
 import MyProfile from "@/components/MyProfile.vue";
 import CreateFlit from "@/components/CreateFlit.vue";
-import FlitFeed from "@/components/FlitFeed.vue";
-import { User } from "@/models/user";
+import UserFlitFeed from "@/components/UserFlitFeed.vue";
+
 import { defineComponent } from "vue";
-import usePosts from "@/composables/usePosts";
 import useUsers from "@/composables/useUsers";
 
 export default defineComponent({
@@ -20,40 +17,15 @@ export default defineComponent({
   components: {
     MyProfile,
     CreateFlit,
-    FlitFeed,
+    UserFlitFeed,
   },
   setup() {
-    const { fetchLoggedUser, loggedUser, fetchUserById, selectedUser } = useUsers()
-    const { posts, limitReached, fetchPostsByUser } = usePosts()
+    const { fetchLoggedUser, loggedUser } = useUsers()
 
     fetchLoggedUser()
 
-    const params = {
-      published: true,
-      page: 1,
-      sort: '-publishedDate',
-      username: loggedUser.value.username
-    }
-    
-    fetchPostsByUser(params)
-
     return {
       loggedUser,
-      posts,
-
-      previousPage: () => {
-        if(params.page < 1) {
-          params.page = params.page - 1;
-          fetchPostsByUser(params)
-        }
-      },
-
-      nextPage: () => {
-        if(!limitReached.value) {
-          params.page ++;
-          fetchPostsByUser(params)
-        }
-      }
     };
   },
 });
